requirements.txt
# Install these requirements if you wish to contribute to this project.

# ===========================================================================
# Packages essential to this app. Needed by anyone who wants to use this app.
# ===========================================================================
django==1.6.2
-e git+https://github.com/divio/django-cms.git@8327844ff84df60e36c9ef1a8f2e63d740519a51#egg=cms
South==0.8.4
django-filer==0.9.5
-e git+https://github.com/KristianOellegaard/django-hvad.git@0e2101f15404eaf9611cd6cf843bfc424117b227#egg=hvad
djangocms-text-ckeditor==2.1
django-countries==1.5
django-multilingual-news==1.3


# ==============================================================
# Packages needed for running the tests. Needed by contributors.
# ==============================================================
Fabric
django-libs
factory-boy
mock
django-nose
coverage
django-coverage
ipython
ipdb
flake8
